move from determined.util to determined.common.utils
