+++ harness/determined/cli/cli.py
@@ -16,7 +16,6 @@ from termcolor import colored
 import determined
 import determined.cli
 import determined.common.api.authentication as auth
-from determined import util
 from determined.cli import checkpoint, experiment, render
 from determined.cli.agent import args_description as agent_args_description
 from determined.cli.master import args_description as master_args_description
@@ -37,7 +36,7 @@ from determined.common import api, yaml
 from determined.common.api.authentication import authentication_required
 from determined.common.check import check_not_none
 from determined.common.declarative_argparse import Arg, Cmd, add_args
-from determined.common.util import chunks, debug_mode, get_default_master_address
+from determined.common.util import chunks, debug_mode, get_default_master_address, safe_load_yaml_with_exceptions
 from determined.deploy.cli import DEPLOY_CMD_NAME
 from determined.deploy.cli import args_description as deploy_args_description
 
@@ -88,7 +87,7 @@ def list_tasks(args: Namespace) -> None:
 
 @authentication_required
 def preview_search(args: Namespace) -> None:
-    experiment_config = util.safe_load_yaml_with_exceptions(args.config_file)
+    experiment_config = safe_load_yaml_with_exceptions(args.config_file)
     args.config_file.close()
 
     if "searcher" not in experiment_config:
