+++ harness/determined/cli/experiment.py
@@ -16,7 +16,7 @@ import tabulate
 import determined.common
 from determined import util
 from determined.cli import checkpoint, render
-from determined.common import api, constants, context, yaml
+from determined.common import api, constants, context, yaml, util
 from determined.common.api.authentication import authentication_required
 from determined.common.declarative_argparse import Arg, Cmd, Group
 from determined.common.experimental import Determined
